feat: checkin-document generation works
