+++ tsconfig.json
@@ -3,8 +3,8 @@
         "outDir": "./dist/",
         "sourceMap": true,
         "noImplicitAny": true,
-        "module": "es6",
-        "target": "es6",
+        "module": "ES2022",
+        "target": "ES2022",
         "jsx": "react",
         "allowJs": true,
         "moduleResolution": "node",
